ci: add Ubuntu 26.04 static analyzer lane
Why:
Keep CI current with the Ubuntu toolchain that developers and future
runners will increasingly see. Ubuntu 26.04 brings newer compiler,
libc, curl, CMake, and clang analyzer behavior, so it is useful as an
early warning lane before broader CI migration.

Impact:
Only the clang static analyzer job moves to Ubuntu 26.04 for now. The
rest of CI stays on the existing images. Follow-up commits are expected
to move additional CI jobs after this canary lane has proven stable.

Before/After:
Before, the static analyzer used the Ubuntu 24.04 dev image. After, it
uses the Ubuntu 26.04 image and the current clang analyzer toolchain.

Technical Overview:
Duplicate the Ubuntu 24.04 dev base image definition for 26.04 and keep
its DockerHub tag as rsyslog/rsyslog_dev_base_ubuntu:26.04.
Use the 24.04 OBS package repository where 26.04 packages are not yet
published, and update bootstrap details for current Ubuntu behavior.
Set the 26.04 container analyzer defaults to the installed clang 21
scan-build tools and llvm-symbolizer.
Switch only the GitHub Actions clang static analyzer job to the new
26.04 image.
Fix the 26.04 image helper script so cleanup does not hide test failure,
and remove a stale sudoers entry for a user that is not created.

Docker Definitions:
Add packaging/docker/dev_env/ubuntu/base/26.04 as the new Ubuntu 26.04
base image directory. It contains the Dockerfile, build helper,
tag-previous helper, and the ClickHouse TLS/config files copied from the
24.04 environment so the new image has the same test support surface.
The Dockerfile remains a direct 24.04 derivative, with only the changes
needed for Ubuntu 26.04 package availability, current tool behavior, and
the clang static analyzer canary.

Static Analyzer Findings:
The newer analyzer found several real defensive gaps. These are fixed
with normal runtime checks and error handling rather than suppressions:
path rewriting now updates the intended byte, empty TCP worksets are
rejected before dereference, template creation errors are propagated,
decrypt reads handle EOF and stream errors cleanly, and statefile,
journal, DTLS socket, and Kubernetes metadata paths now guard invalid
or missing values.

The newer Ubuntu headers also exposed portability issues in libcurl
calls. Those were not analyzer bugs, but the 26.04 toolchain correctly
requires long or size-typed arguments for several curl options. The
affected HTTP, ClickHouse, Elasticsearch, Kubernetes, HTTPFS, and fmhttp
paths now pass the expected types.

One report led to hygiene rather than a proven runtime bug: omfwd now
stores errno immediately after sendto so later logging cannot obscure
the original send failure.

The remaining analyzer reports were reviewed as false positives and are
suppressed narrowly for clang analyzer builds only. KSI state restore
and IV seeding are serialized by module_lock, and omprog status reads
are bounded by the preceding poll timeout. The suppression sites include
inline comments explaining those assumptions.

Docker image:
rsyslog/rsyslog_dev_base_ubuntu:26.04
sha256:576d27bf5c9c783989caf8aeeae0c6bce1434b547b09fdf1ae956033058144f5

Validation:
devtools/format-code.sh --git-changed
Docker image rebuild and post-build tool/user checks.
sh -n packaging/docker/dev_env/ubuntu/base/26.04/build.sh
git diff --check
static analyzer in rsyslog/rsyslog_dev_base_ubuntu:26.04
with SCAN_BUILD=scan-build, SCAN_BUILD_CC=clang, and
SCAN_BUILD_REPORT_DIR=scan-build-report.
The run starts from make clean inside the dev container and reports:
scan-build: No bugs found.
static analyzer result: 0

With the help of AI-Agents: Codex, Rawls

+++ contrib/omhttpfs/omhttpfs.c
@@ -451,7 +451,7 @@ replication - required block replication for the file.
     httpfs_set_url(pWrkrData, "&op=create&overwrite=false&data=true");
 
     curl_easy_setopt(pWrkrData->curl, CURLOPT_POSTFIELDS, (char*)buf);
-    curl_easy_setopt(pWrkrData->curl, CURLOPT_POSTFIELDSIZE, strlen((char*)buf));
+    curl_easy_setopt(pWrkrData->curl, CURLOPT_POSTFIELDSIZE, (long)strlen((char*)buf));
 
     DBGPRINTF("%s(): msg=%s\n", __FUNCTION__, buf);
 
@@ -493,7 +493,7 @@ static rsRetVal httpfs_append_file(wrkrInstanceData_t* pWrkrData, uchar* buf) {
     httpfs_set_url(pWrkrData, "&op=append&data=true");
 
     curl_easy_setopt(pWrkrData->curl, CURLOPT_POSTFIELDS, (char*)buf);
-    curl_easy_setopt(pWrkrData->curl, CURLOPT_POSTFIELDSIZE, strlen((char*)buf));
+    curl_easy_setopt(pWrkrData->curl, CURLOPT_POSTFIELDSIZE, (long)strlen((char*)buf));
 
     headers = httpfs_curl_add_header(headers, 1, HTTPFS_CONTENT_TYPE);
     curl_easy_setopt(pWrkrData->curl, CURLOPT_HTTPHEADER, headers);
